# VmCloudController spec speedup.