Before:
```
Finished in 31.25 seconds (files took 9.39 seconds to load)                         
133 examples, 0 failures  
```

After:
```
Finished in 20.98 seconds (files took 9.53 seconds to load)                              
133 examples, 0 failures  
```